rate limit display on RecentlyActiveUsers
