+++ packages/lesswrong/lib/index.ts
@@ -257,3 +257,6 @@ import './vulcan-i18n-en-us';
 import '../components/spuriousChange';
 
 import './sql';
+
+import './rateLimits/constants';
+import './rateLimits/types';
