+++ packages/lesswrong/lib/rateLimits/constants.ts
@@ -1,26 +1,29 @@
-import type { ForumOptions } from "../../lib/forumTypeUtils";
-import { RECENT_CONTENT_COUNT } from "../repos/VotesRepo";
+import type { ForumOptions } from "../forumTypeUtils";
 import type { AutoRateLimit, CommentAutoRateLimit, PostAutoRateLimit, TimeframeUnitType } from "./types";
 
 /**
  * Post rate limits
  */
 
- type TimeframeStrType<A extends AutoRateLimit['actionType']> = `${number} ${A} per ${number} ${TimeframeUnitType}` 
+export const RECENT_CONTENT_COUNT = 20
 
- const timeframe = <A extends AutoRateLimit['actionType']>(timeframeString: TimeframeStrType<A>) => {
-   const [itemsPerTimeframe, actionType, _, timeframeLength, timeframeUnit] = timeframeString.split(' ');
-   return {
-     itemsPerTimeframe: parseInt(itemsPerTimeframe),
-     actionType: actionType as A,
-     timeframeLength: parseInt(timeframeLength),
-     timeframeUnit: timeframeUnit as TimeframeUnitType
-   }
- };
+type TimeframeStrType<A extends AutoRateLimit['actionType']> = `${number} ${A} per ${number} ${TimeframeUnitType}` 
 
-const EA = {
-  POSTS: {
-  },
+// It felt pretty important to Ray to easily skim all the rate limits, so Robert worked out this 
+// parser to make the basic settings more human-readable
+const timeframe = <A extends AutoRateLimit['actionType']>(timeframeString: TimeframeStrType<A>) => {
+  const [itemsPerTimeframe, actionType, _, timeframeLength, timeframeUnit] = timeframeString.split(' ');
+  return {
+    itemsPerTimeframe: parseInt(itemsPerTimeframe),
+    actionType: actionType as A,
+    timeframeLength: parseInt(timeframeLength),
+    timeframeUnit: timeframeUnit as TimeframeUnitType
+  }
+};
+
+const EA: {POSTS: PostAutoRateLimit[], COMMENTS: CommentAutoRateLimit[]} = {
+  POSTS: [
+  ],
   COMMENTS: [
     {
       ...timeframe('4 Comments per 30 minutes'),
@@ -30,15 +33,15 @@ const EA = {
       appliesToOwnPosts: true
     }, {
       ...timeframe('4 Comments per 30 minutes'),
-      downvoteRatio: .3,
+      downvoteRatioThreshold: .3,
       rateLimitType: "downvoteRatio",
       rateLimitMessage: "",
       appliesToOwnPosts: true
     },
   ]
-} as const
+}
 
-const LW = {
+const LW: {POSTS: PostAutoRateLimit[], COMMENTS: CommentAutoRateLimit[]} = {
   POSTS: [
     {
       ...timeframe('2 Posts per 1 weeks'),
@@ -68,22 +71,17 @@ const LW = {
   ],
   COMMENTS: [ 
     {
+      ...timeframe('1 Comments per 1 hours'),
+      recentKarmaThreshold: -1,
+      downvoterCountThreshold: 3,
+      appliesToOwnPosts: true,
+      rateLimitMessage: `Users are limited to 3 comments/day unless their last ${RECENT_CONTENT_COUNT} posts/comments have at least 1 net-karma.`
+    }, {
       ...timeframe('3 Comments per 1 days'),
       karmaThreshold: 5,
       appliesToOwnPosts: true,
       rateLimitMessage: "New users can write up to 3 comments a day. Gain more karma to comment more frequently.",
     }, {
-      ...timeframe('1 Comments per 1 days'),
-      karmaThreshold: -1,
-      appliesToOwnPosts: false,
-      rateLimitMessage: "Negative karma users are limited to 1 comment per day.",
-    }, {
-      ...timeframe('1 Comments per 3 days'),
-      karmaThreshold: -15,
-      appliesToOwnPosts: false,
-      rateLimitMessage: "Users with -15 or less karma users can only comment once per 3 days.",
-    }, 
-    {
       ...timeframe('3 Comments per 1 days'),
       // semi-established users can make up to 20 posts/comments without getting upvoted, before hitting a 3/day comment rate limit
       karmaThreshold: 1999, // at 2000+ karma I trust you more to go on long conversations
@@ -91,11 +89,15 @@ const LW = {
       appliesToOwnPosts: false,
       rateLimitMessage: `Hey, you've been posting a lot without getting upvoted. Users are limited to 3 comments/day unless their last ${RECENT_CONTENT_COUNT} posts/comments have at least 2 net-karma.`,
     }, {
-      ...timeframe('1 Comments per 1 hours'),
-      recentKarmaThreshold: -1,
-      downvoterCountThreshold: 3,
-      appliesToOwnPosts: true,
-      rateLimitMessage: `Users are limited to 3 comments/day unless their last ${RECENT_CONTENT_COUNT} posts/comments have at least 1 net-karma.`
+      ...timeframe('1 Comments per 1 days'),
+      karmaThreshold: -1,
+      appliesToOwnPosts: false,
+    rateLimitMessage: "Negative karma users are limited to 1 comment per day.",
+    }, {
+      ...timeframe('1 Comments per 3 days'),
+      karmaThreshold: -15,
+      appliesToOwnPosts: false,
+      rateLimitMessage: "Users with -15 or less karma users can only comment once per 3 days.",
     }, {
       ...timeframe('1 Comments per 1 days'),
       recentKarmaThreshold: -5,
@@ -107,17 +109,17 @@ const LW = {
       recentKarmaThreshold: -15,
       downvoterCountThreshold: 2,
       appliesToOwnPosts: false,
-      rateLimitMessage: `Your recent posts/comments have been net-downvoted. Users with -60 or less karma on their recent ${RECENT_CONTENT_COUNT} posts/comments can only comment once per two weeks on other's posts.`,
+      rateLimitMessage: `Your recent posts/comments have been net-downvoted. Users with -15 or less net-karma on their recent ${RECENT_CONTENT_COUNT} posts/comments can only comment once per two weeks on other's posts.`,
     }, {
       ...timeframe('1 Comments per 1 weeks'),
       karmaThreshold: 1999, // at 2000+ karma I trust you more to go on long conversations
-      lastMonthKarma: -30,
+      lastMonthKarmaThreshold: -30,
       downvoterCountThreshold: 2,
       appliesToOwnPosts: false,
-      rateLimitMessage: `Your recent posts/comments have been net-downvoted. Users with -60 or less karma on their recent ${RECENT_CONTENT_COUNT} posts/comments can only comment once per two weeks on other's posts.`
+      rateLimitMessage: `Your recent posts/comments have been net-downvoted. Users with -30 or less net-karma on their recent ${RECENT_CONTENT_COUNT} posts/comments can only comment once per two weeks on other's posts.`
     }
   ]
-} as const
+}
 
 const ALL = {
   POSTS: {
